Changeset: 59 Upgrade Unity to 2021.3.23f1, also upgrade many packages
WebXR-Export breaks in this version because com.unity.subsystemregistration was deprecated and
removed as a dependency of com.unity.xr.management. Re-adding com.unity.subsystemregistration
as a top-level dependency fixes it while we wait for DePantehr to migrate away from the package,
as long as Unity doesn't remove it before then.